Refactor SPIR-V IR into separate header / source files.
