+++ dependencies/CMakeLists.txt
@@ -25,6 +25,7 @@ add_subdirectory(llvm)
 
 add_subdirectory(jpeg)
 add_subdirectory(png)
+add_subdirectory(spirv)
 
 # Needs cache vars set by llvm, do not reorder.
 add_subdirectory(wasm)
